+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fixes issue where upper case file names would not be matched in `parse` 
 - Reduce batch size when deleting comment batches
 - Support attachment type filters
+- support getting stats for `get buckets` 
+
 
 # v0.24.0
 - BREAKING: the `--context` option is now required. Users need to opt